elt_doc_transcript_audio/pyproject.toml
[project]
name = "elt-doc-transcript-audio"
version = "0.1.0"
description = "ELT module for transcribing audio recordings and extracting meeting discussions"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai-whisper>=20250625",
]

[build-system]
requires = ["uv_build>=0.9.15,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
